refactor(debug): consolidate debug modes, fields and value decode into one source
Make src/js/utils/debugModes.js the single source of truth for all
per-debug-mode knowledge — mode list, field labels, value→friendly
formatting and value↔chart conversion — consumed by both the sensors
tab live view (via the debug store) and the blackbox viewer.

- Add getDebugFieldNames(apiVersion) and the pure decodeDebugFieldToFriendly /
  convertDebugFieldValue helpers (hardware scaling injected via a ctx object);
  slim stores/debug.js to thin wrappers.
- Blackbox viewer: drop its parallel DEBUG_MODE / DEBUG_FRIENDLY_FIELD_NAMES
  tables and adjustDebugDefsList; the presenter keeps thin wrappers that build
  the scaling ctx from the parsed log.
- Resolve the log's API version once at the parser core
  (sysConfig.apiVersion), mirroring the FC store's config.apiVersion;
  consumers read it directly instead of mapping per call.
- Wire the sensors live view to render decoded debug values with units.
