+++ src/blackbox-viewer/values_display.js
@@ -51,12 +51,14 @@ export function updateValuesChart(logStore, graphStore, appStore, userSettings) 
     const currentFlightMode = frame[logStore.flightLog.getMainFieldIndexByName("flightModeFlags")];
 
     if (graphStore.hasTableOverlay) {
-        const debugMode = logStore.flightLog.getSysConfig().debug_mode;
+        const sysConfig = logStore.flightLog.getSysConfig();
+        const debugMode = sysConfig.debug_mode;
+        const apiVersion = sysConfig.apiVersion;
         const values = [];
 
         for (let i = 0; i < fieldNames.length; i++) {
             values.push({
-                name: FlightLogFieldPresenter.fieldNameToFriendly(fieldNames[i], debugMode),
+                name: FlightLogFieldPresenter.fieldNameToFriendly(fieldNames[i], debugMode, apiVersion),
                 raw: atMost2DecPlaces(frame[i]),
                 decoded: FlightLogFieldPresenter.decodeFieldToFriendly(
                     logStore.flightLog,
@@ -76,7 +78,7 @@ export function updateValuesChart(logStore, graphStore, appStore, userSettings) 
                 continue;
             }
             statRows.push({
-                name: FlightLogFieldPresenter.fieldNameToFriendly(stat.name, debugMode),
+                name: FlightLogFieldPresenter.fieldNameToFriendly(stat.name, debugMode, apiVersion),
                 min: `${FlightLogFieldPresenter.decodeFieldToFriendly(logStore.flightLog, stat.name, stat.min)} (${atMost2DecPlaces(stat.min)})`,
                 max: `${FlightLogFieldPresenter.decodeFieldToFriendly(logStore.flightLog, stat.name, stat.max)} (${atMost2DecPlaces(stat.max)})`,
                 mean: `${FlightLogFieldPresenter.decodeFieldToFriendly(logStore.flightLog, stat.name, stat.mean)} (${atMost2DecPlaces(stat.mean)})`,
